refactor: move PokeDicText to component package and clean up imports
