+++ content/presentation/src/main/java/com/example/content/presentation/details/component/PokemonDetailltem.kt
@@ -8,7 +8,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.content.presentation.details.libs.PokeDicText
 import com.example.core.presentation.designsystem.PokeDicTheme
 
 @Composable
